+++ src/main/java/com/beansgalaxy/beansbackpacks/item/BackpackItem.java
@@ -3,16 +3,13 @@ package com.beansgalaxy.beansbackpacks.item;
 import com.beansgalaxy.beansbackpacks.entity.BackpackEntity;
 import com.beansgalaxy.beansbackpacks.entity.Kind;
 import com.beansgalaxy.beansbackpacks.entity.PlaySound;
-import com.beansgalaxy.beansbackpacks.networking.packages.SyncBackSlot;
+import com.beansgalaxy.beansbackpacks.networking.server.sSyncBackSlot;
 import com.beansgalaxy.beansbackpacks.screen.BackSlot;
-import net.minecraft.client.MinecraftClient;
-import net.minecraft.client.item.BundleTooltipData;
 import net.minecraft.client.item.TooltipData;
 import net.minecraft.entity.player.PlayerEntity;
 import net.minecraft.item.*;
 import net.minecraft.nbt.NbtCompound;
 import net.minecraft.nbt.NbtElement;
-import net.minecraft.screen.slot.Slot;
 import net.minecraft.server.network.ServerPlayerEntity;
 import net.minecraft.util.ActionResult;
 import net.minecraft.util.collection.DefaultedList;
@@ -53,7 +50,7 @@ public class BackpackItem extends Item implements DyeableItem {
         Box box = BackpackEntity.newBox(blockPos, y, 10 / 16d, direction);
         if (player.getWorld().isSpaceEmpty(box) && BackpackItem.doesPlace(player, x, y, z, direction, backpackStack, true)) {
             if (player instanceof ServerPlayerEntity serverPlayer)
-                SyncBackSlot.S2C(serverPlayer);
+                sSyncBackSlot.S2C(serverPlayer);
             return ActionResult.SUCCESS;
         }
 
@@ -74,7 +71,7 @@ public class BackpackItem extends Item implements DyeableItem {
         boolean spaceEmpty = player.getWorld().isSpaceEmpty(box);
         if (spaceEmpty && doesPlace(player, x, y, z, direction, backpackStack, copyBackpackInventory)) {
             if (player instanceof ServerPlayerEntity serverPlayer)
-                SyncBackSlot.S2C(serverPlayer);
+                sSyncBackSlot.S2C(serverPlayer);
             return ActionResult.SUCCESS;
         }
 
@@ -125,46 +122,12 @@ public class BackpackItem extends Item implements DyeableItem {
         return YRot;
     }
 
-    private static int getItemOccupancy(ItemStack stack) {
-        NbtCompound nbtCompound;
-        if ((stack.isOf(Items.BEEHIVE) || stack.isOf(Items.BEE_NEST)) && stack.hasNbt() && (nbtCompound = BlockItem.getBlockEntityNbt(stack)) != null && !nbtCompound.getList("Bees", NbtElement.COMPOUND_TYPE).isEmpty()) {
-            return 64;
-        }
-        return 64 / stack.getMaxCount();
-    }
-
-    private static int getBundleOccupancy(DefaultedList<ItemStack> defaultedList) {
-        return defaultedList.stream().mapToInt(itemStack -> BackpackItem.getItemOccupancy(itemStack) * itemStack.getCount()).sum();
-    }
-
     // GIVES AN EQUIPPED BACKPACK A CUSTOM TOOLTIP.
     public Optional<TooltipData> getTooltipData(ItemStack stack) {
-        PlayerEntity player = MinecraftClient.getInstance().player;
-        DefaultedList<Slot> slots = MinecraftClient.getInstance().player.playerScreenHandler.slots;
-        // IS BACKPACK SLOT LOADED AND PLAYER IS LOOKING IN THE PLAYER INVENTORY
-        if (slots.size() <= BackSlot.SLOT_INDEX || player.currentScreenHandler != player.playerScreenHandler)
+        Optional<TooltipData> get = Tooltip.get(stack);
+        if (get == null)
             return super.getTooltipData(stack);
-        ItemStack equippedOnBack = slots.get(BackSlot.SLOT_INDEX).getStack();
-        // IS BACKPACK EQUIPPED
-        if (!stack.equals(equippedOnBack))
-            return super.getTooltipData(stack);
-        DefaultedList<ItemStack> defaultedList = DefaultedList.of();
-        DefaultedList<ItemStack> backpackList = BackSlot.getInventory(player).getItemStacks();
-        backpackList.forEach(itemstack -> defaultedList.add(itemstack.copy()));
-        if (!defaultedList.isEmpty()) {
-            defaultedList.remove(0);
-            for (int j = 0; j < defaultedList.size(); j++) {
-                ItemStack itemStack = defaultedList.get(j);
-                int count = defaultedList.stream()
-                        .filter(itemStacks -> itemStacks.isOf(itemStack.getItem()) && !itemStack.equals(itemStacks))
-                        .mapToInt(itemStacks -> itemStacks.copyAndEmpty().getCount()).sum();
-                itemStack.setCount(count + itemStack.getCount());
-                defaultedList.removeIf(ItemStack::isEmpty);
-            }
-
-        }
-        int totalWeight = BackpackItem.getBundleOccupancy(defaultedList) / Kind.getMaxStacks(equippedOnBack);
-        return Optional.of(new BundleTooltipData(defaultedList, totalWeight));
+        return get;
     }
 
 }