+++ src/main/java/com/beansgalaxy/beansbackpacks/networking/client/SyncBackpackInventory.java
@@ -1,0 +1,4 @@
+package com.beansgalaxy.beansbackpacks.networking.client;
+
+public class SyncBackpackInventory {
+}