+++ src/main/java/com/beansgalaxy/beansbackpacks/networking/packages/SprintKeyPacket.java
@@ -4,14 +4,10 @@ import com.beansgalaxy.beansbackpacks.networking.NetworkPackages;
 import com.beansgalaxy.beansbackpacks.screen.BackSlot;
 import net.fabricmc.fabric.api.client.networking.v1.ClientPlayNetworking;
 import net.fabricmc.fabric.api.networking.v1.PacketByteBufs;
-import net.fabricmc.fabric.api.networking.v1.PacketSender;
 import net.minecraft.client.MinecraftClient;
 import net.minecraft.client.network.ClientPlayerEntity;
 import net.minecraft.client.util.InputUtil;
 import net.minecraft.network.PacketByteBuf;
-import net.minecraft.server.MinecraftServer;
-import net.minecraft.server.network.ServerPlayNetworkHandler;
-import net.minecraft.server.network.ServerPlayerEntity;
 
 public class SprintKeyPacket {
     public static void updateSprintKeyIsPressed(ClientPlayerEntity clientPlayer) {
@@ -35,8 +31,4 @@ public class SprintKeyPacket {
         ClientPlayNetworking.send(NetworkPackages.SPRINT_KEY_2S, buf);
     }
 
-    public static void receiveAtServer(MinecraftServer server, ServerPlayerEntity serverPlayer, ServerPlayNetworkHandler handler,
-                                       PacketByteBuf buf, PacketSender responseSender) {
-        BackSlot.get(serverPlayer).sprintKeyIsPressed = buf.readBoolean();
-    }
 }